Conflicts:
	src/main/java/br/jus/stf/plataforma/shared/security/SecurityChecker.java

+++ src/main/java/br/jus/stf/plataforma/pesquisas/infra/configuration/ElasticsearchLocalAfterPropertiesSetConfigurer.java
@@ -8,9 +8,12 @@ import org.elasticsearch.common.unit.TimeValue;
 import org.elasticsearch.env.Environment;
 import org.elasticsearch.plugins.PluginManager;
 import org.elasticsearch.plugins.PluginManager.OutputMode;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.elasticsearch.ElasticsearchProperties;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
+import br.jus.stf.plataforma.shared.persistence.LocalData;
 import br.jus.stf.plataforma.shared.settings.Profiles;
 
 /**
@@ -24,20 +27,33 @@ import br.jus.stf.plataforma.shared.settings.Profiles;
 public class ElasticsearchLocalAfterPropertiesSetConfigurer implements ElasticsearchAfterPropertiesSetConfigurer {
 
 	private static final String ELASTIC_DATA_DIR = "target/elasticsearch/data";
+	
+	@Autowired
+	private org.springframework.core.env.Environment env;
+	
+	@Autowired
+	private ElasticsearchProperties elasticsearchProperties;
 
 	@Override
 	public void afterPropertiesSet() {
-		// Apaga o diretório de dados.
-		FileSystemUtils.deleteRecursively(new File(ELASTIC_DATA_DIR), true);
+		if (env.acceptsProfiles(Profiles.KEEP_DATA)) {
+			// Para manter os dados, altera o diretório em que será armazenado o elasticsearch.
+			elasticsearchProperties.getProperties().put("path.data", LocalData.instance().dataDirAbsolutePath() + "/elasticsearch/data");
+			elasticsearchProperties.getProperties().put("path.logs", LocalData.instance().dataDirAbsolutePath() + "/elasticsearch/logs");
+			elasticsearchProperties.getProperties().put("path.plugins", LocalData.instance().dataDirAbsolutePath() + "/elasticsearch/plugins");
+		} else {
+			// Apaga o diretório de dados.
+			FileSystemUtils.deleteRecursively(new File(ELASTIC_DATA_DIR), true);
+		}
 	}
 
 	@Override
 	public void afterPropertiesSet(Client elasticsearchClient) throws Exception {
-		Environment env = new Environment(elasticsearchClient.settings());
+		Environment envir = new Environment(elasticsearchClient.settings());
 
-		if (!env.pluginsFile().exists()) {
-			FileSystemUtils.mkdirs(env.pluginsFile());
-			PluginManager pluginManager = new PluginManager(env, null, OutputMode.DEFAULT,
+		if (!envir.pluginsFile().exists()) {
+			FileSystemUtils.mkdirs(envir.pluginsFile());
+			PluginManager pluginManager = new PluginManager(envir, null, OutputMode.DEFAULT,
 					TimeValue.timeValueMinutes(2));
 			pluginManager.downloadAndExtract("mobz/elasticsearch-head");
 		}
